requirements.txt
requests>=2.31.0
urwid>=4.0.0
python-dotenv>=1.2.2